seperate editarea from editor to allow function redefinitions
